Module potatosync_rust::account::model

Models for handling Accounts

Structs

Account
General Account struct used for retrieving from DB and updating

Image
Used when changing Profile picture

InfoResponse
Struct used when returning account info to client

LoginCredentials
Used for getting the credentials from the client

NewAccount
Used for verifying if the entered info when registering is valid

Password
Used when changing password

PatchingAccount

TokenAccount
Struct used for returning account info in TokenResponse

TokenResponse
Struct used for sending back account info when you have logged in

Username
Used when changing username

Functions

_impl_as_changeset_for_account

_impl_insertable_for_account

_impl_queryable_for_account

deserialize_option
Special function to deserialize Option<DateTime>

serialize_option
Special function to serialize Option<DateTime>

validate_email
Checks if email is valid and doesnt exist in DB

validate_password
Checks if password is valid

validate_username
Checks if username is valid and doesnt exist in DB