Module potatosync_rust::note::model

Models used for handling notes

Structs

Note
General Note struct used for retrieving from db and updating notes

NoteId
Struct used when client needs to specify certain note_id

NoteLastUpdated
Struct used when client needs to specify last_updated timestamp

NoteResponse
Struct used when server needs to return list of notes along with status of request

PatchingNote

SavingNote
Note as provided by the client when saving. Note the missing account_id since the client doesnt know the id.

Functions

_impl_as_changeset_for_note

_impl_as_changeset_for_patchingnote

_impl_insertable_for_note

_impl_insertable_for_patchingnote

_impl_insertable_for_savingnote

_impl_queryable_for_note

_impl_queryable_for_patchingnote

_impl_queryable_for_savingnote